+++ packages/ui-react/src/components/modal/ModalStyles.tsx
@@ -3,6 +3,8 @@ import styled, { css } from '../../utils/styledWrapper.js';
 import { Close, ArrowBack } from '../../icons';
 import { ButtonIcon } from '../button';
 
+const MaxInnerWidth = 600;
+
 export const ModalStyle = styled.div<{ $center: boolean }>`
   ${({
     theme: { fontSizesMap, borderRadiusesMap, colors, boxShadows },
@@ -33,9 +35,10 @@ export const ModalHeaderStyle = styled.div<{
   $short: boolean;
 }>`
   ${({ theme: { spaceMap, fontSizesMap, mediaQueries }, $short }) => css`
+    max-width: ${MaxInnerWidth}px;
+    min-height: 32px;
     display: flex;
     align-items: flex-start;
-    min-height: 32px;
     margin-bottom: ${$short ? -spaceMap.md : 0}px;
     padding: ${spaceMap.xl}px ${spaceMap.xxl}px;
     font-size: ${fontSizesMap.md}px;
@@ -100,6 +103,7 @@ export const ModalSubtitleStyle = styled.div`
     line-height: 24px;
     margin-top: -${spaceMap.xl}px;
     padding: 0 ${spaceMap.xxl}px ${spaceMap.sm}px;
+    max-width: ${MaxInnerWidth}px;
 
     ${mediaQueries.md} {
       padding-left: ${spaceMap.xl}px;
@@ -110,6 +114,7 @@ export const ModalSubtitleStyle = styled.div`
 
 export const ModalContentStyle = styled.div`
   ${({ theme: { spaceMap, mediaQueries } }) => css`
+    max-width: ${MaxInnerWidth}px;
     padding: 0 ${spaceMap.xxl}px 0;
 
     ${mediaQueries.md} {