+++ packages/wallets/exodus/src/index.ts
@@ -1,5 +1,15 @@
-import { WalletAdapterType } from '@reef-knot/core-react';
+import { WalletAdapterType } from '@reef-knot/types';
+import { Ethereum } from '@wagmi/core';
+import { InjectedConnector } from 'wagmi/connectors/injected';
 import WalletIcon from './icons/exodus.svg';
+
+declare global {
+  interface Window {
+    // @ts-expect-error Subsequent property declarations must have the same type
+    ethereum?: Ethereum & { isExodus?: boolean };
+    exodus?: { ethereum?: Ethereum & { isExodus?: boolean } };
+  }
+}
 
 export const Exodus: WalletAdapterType = () => ({
   walletName: 'Exodus',
@@ -8,9 +18,20 @@ export const Exodus: WalletAdapterType = () => ({
     light: WalletIcon,
     dark: WalletIcon,
   },
-  detector: () => !!window.ethereum?.isExodus,
+  detector: () =>
+    typeof window !== 'undefined'
+      ? !!window.exodus?.ethereum?.isExodus || !!window.ethereum?.isExodus
+      : false,
   downloadURLs: {
     default: 'https://www.exodus.com/download/',
   },
-  // TODO: handle wallet conflicts and custom connectors
+  connector: new InjectedConnector({
+    options: {
+      name: 'Exodus',
+      getProvider: () =>
+        typeof window !== 'undefined'
+          ? window.exodus?.ethereum || window.ethereum
+          : undefined,
+    },
+  }),
 });