+++ packages/web3-react/test/hooks/useConnectorImToken.test.ts
@@ -2,6 +2,7 @@ jest.mock('tiny-warning');
 jest.mock('../../src/helpers/openWindow');
 jest.mock('../../src/hooks/useWeb3');
 jest.mock('../../src/hooks/useConnectors');
+jest.mock('wagmi');
 
 const mockIsMobileOrTablet = jest.fn();
 jest.mock('../../src/helpers/ua', () => ({
@@ -10,6 +11,7 @@ jest.mock('../../src/helpers/ua', () => ({
   },
 }));
 
+import { useDisconnect } from 'wagmi';
 import warning from 'tiny-warning';
 import { renderHook, act } from '@testing-library/react-hooks';
 import { openWindow } from '../../src/helpers/openWindow';
@@ -23,11 +25,14 @@ const mockUseConnectors = useConnectors as jest.MockedFunction<
 >;
 const mockOpenWindow = openWindow as jest.MockedFunction<typeof openWindow>;
 const mockWarning = warning as jest.MockedFunction<typeof warning>;
+const mockUseDisconnect = useDisconnect as jest.MockedFunction<any>;
 
 beforeEach(() => {
+  const mockDisconnect = jest.fn(async () => true);
   delete window.ethereum;
   mockUseWeb3.mockReturnValue({} as any);
   mockUseConnectors.mockReturnValue({ injected: {} } as any);
+  mockUseDisconnect.mockReturnValue({ disconnectAsync: mockDisconnect });
   mockOpenWindow.mockReset();
   mockWarning.mockReset();
   mockIsMobileOrTablet.mockReturnValue(true);
@@ -44,11 +49,12 @@ describe('useConnectorImToken', () => {
 
   test('should connect if ethereum and imToken are presented', async () => {
     const mockActivate = jest.fn(async () => true);
+    const mockDeactivate = jest.fn(async () => true);
     const injected = {};
 
     window.ethereum = {};
     window.ethereum.isImToken = true;
-    mockUseWeb3.mockReturnValue({ activate: mockActivate } as any);
+    mockUseWeb3.mockReturnValue({ activate: mockActivate, deactivate: mockDeactivate } as any);
     mockUseConnectors.mockReturnValue({ injected } as any);
 
     const { result } = renderHook(() => useConnectorImToken());