+++ packages/web3-react/test/hooks/useConnectorTally.test.ts
@@ -1,7 +1,9 @@
 jest.mock('../../src/helpers/openWindow');
 jest.mock('../../src/hooks/useWeb3');
 jest.mock('../../src/hooks/useConnectors');
+jest.mock('wagmi');
 
+import { useDisconnect } from 'wagmi';
 import { renderHook, act } from '@testing-library/react-hooks';
 import { openWindow } from '../../src/helpers/openWindow';
 import { useConnectorTally } from '../../src/hooks/useConnectorTally';
@@ -13,22 +15,26 @@ const mockUseConnectors = useConnectors as jest.MockedFunction<
   typeof useConnectors
 >;
 const mockOpenWindow = openWindow as jest.MockedFunction<typeof openWindow>;
+const mockUseDisconnect = useDisconnect as jest.MockedFunction<any>;
 
 beforeEach(() => {
+  const mockDisconnect = jest.fn(async () => true);
   delete window.ethereum;
   mockUseWeb3.mockReturnValue({} as any);
   mockUseConnectors.mockReturnValue({ injected: {} } as any);
+  mockUseDisconnect.mockReturnValue({ disconnectAsync: mockDisconnect });
   mockOpenWindow.mockReset();
 });
 
 describe('useConnectorTally', () => {
   test('should connect if ethereum and Tally are presented', async () => {
     const mockActivate = jest.fn(async () => true);
+    const mockDeactivate = jest.fn(async () => true);
     const injected = {};
 
     window.ethereum = {};
     window.ethereum.isTally = true;
-    mockUseWeb3.mockReturnValue({ activate: mockActivate } as any);
+    mockUseWeb3.mockReturnValue({ activate: mockActivate, deactivate: mockDeactivate } as any);
     mockUseConnectors.mockReturnValue({ injected } as any);
 
     const { result } = renderHook(() => useConnectorTally());