+++ packages/connect-wallet-modal/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @reef-knot/connect-wallet-modal
 
+## 1.0.5
+
+### Patch Changes
+
+- Support custom connectors for wallet adapters. Improve getWalletsButtons().
+- Updated dependencies
+  - @reef-knot/core-react@1.0.1
+  - @reef-knot/web3-react@1.0.4
+  - @reef-knot/ui-react@1.0.2
+
 ## 1.0.4
 
 ### Patch Changes